// src/app/models/customer.ts
export interface Customer{
    firstName: string;
    lastName: string;
    email?: string;
    address?: string;
    phone?: string;
    gender?: string;
    image?: string;
    id?: number;
    userCreated?: number;
    userModified?: number;
    dateCreated?: string;
    dateModified?: string;
    fullName?: string;
}